Calibration Error Solved During AC Current Flow throw the Sensor
calibrate() method calibrates zero point of sensor,
It is not necessary, but may positively affect the accuracy. Our line frequency is 50Hz where the time period is about 20ms. So if we apply an operation on a single cycle then we get the actual zero point during ac current is flowed. And here the solution by sampling a single cycle .